Fix splash fade animation.
