+++ frontend-v2/src/components/Splash.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="splash" :class="[show ? 'visible' : 'hidden']" @click="hide">
+  <div class="splash">
     <div>
       <img class="logo" src="../assets/holoholo-logo@3x.png" alt="Holo Holo">
     </div>
@@ -12,27 +12,13 @@ import RibbonComponent from '@/components/Ribbon'
 
 export default {
   name: 'Splash',
-  // data: () => ({isActive: true}),
-  props: {
-    show: false
-  },
   components: {
     RibbonComponent
-  },
-  mounted: function () {
-    setTimeout(() => {
-      this.hide()
-    }, 3000)
-  },
-  methods: {
-    hide: function () {
-      this.isActive = false
-    }
   }
 }
 </script>
 
-<style>
+<style scoped>
 .splash {
   background: url("../assets/splash-desktop.jpg") no-repeat center center fixed;
   background-size: cover;
@@ -46,17 +32,14 @@ export default {
   bottom: 0;
   left: 0;
   z-index: 99;
-  transition: opacity 0.5s ease;
 }
 
-.visible {
-  opacity: 1;
-  pointer-events: all;
+.fade-enter-active, .fade-leave-active {
+  transition: opacity .5s;
 }
 
-.hidden {
+.fade-enter, .fade-leave-to {
   opacity: 0;
-  pointer-events: none;
 }
 
 @media (min-width: 320px) and (max-width: 767px) {
